requirements.txt
pyTooling[terminal] ~= 6.5
ruamel.yaml ~= 0.18.6
